+++ apiserver/controller/src/main/java/com/zipline/controller/customer/CustomerController.java
@@ -15,16 +15,16 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.zipline.dto.CustomerDetailResponseDTO;
-import com.zipline.dto.CustomerListResponseDTO;
-import com.zipline.dto.CustomerModifyRequestDTO;
-import com.zipline.dto.CustomerRegisterRequestDTO;
-import com.zipline.dto.PageRequestDTO;
-import com.zipline.dto.agentProperty.AgentPropertyListResponseDTO;
-import com.zipline.dto.contract.ContractListResponseDTO;
-import com.zipline.dto.counsel.CounselCreateRequestDTO;
-import com.zipline.dto.counsel.CounselListResponseDTO;
+import com.zipline.service.customer.dto.response.CustomerDetailResponseDTO;
+import com.zipline.service.customer.dto.response.CustomerListResponseDTO;
+import com.zipline.service.customer.dto.request.CustomerModifyRequestDTO;
+import com.zipline.service.customer.dto.request.CustomerRegisterRequestDTO;
+import com.zipline.global.request.PageRequestDTO;
+import com.zipline.service.contract.dto.response.ContractListResponseDTO;
+import com.zipline.service.counsel.dto.request.CounselCreateRequestDTO;
+import com.zipline.service.counsel.dto.response.CounselListResponseDTO;
 import com.zipline.global.response.ApiResponse;
+import com.zipline.service.agentProperty.dto.response.AgentPropertyListResponseDTO;
 import com.zipline.service.counsel.CounselService;
 import com.zipline.service.customer.CustomerService;
