+++ apiserver/service/src/main/java/com/zipline/service/customer/CustomerServiceImpl.java
@@ -8,14 +8,6 @@ import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.zipline.dto.CustomerDetailResponseDTO;
-import com.zipline.dto.CustomerListResponseDTO;
-import com.zipline.dto.CustomerModifyRequestDTO;
-import com.zipline.dto.CustomerRegisterRequestDTO;
-import com.zipline.dto.PageRequestDTO;
-import com.zipline.dto.agentProperty.AgentPropertyListResponseDTO;
-import com.zipline.dto.contract.ContractListResponseDTO;
-import com.zipline.dto.counsel.CounselListResponseDTO;
 import com.zipline.entity.agentProperty.AgentProperty;
 import com.zipline.entity.contract.CustomerContract;
 import com.zipline.entity.counsel.Counsel;
@@ -24,12 +16,20 @@ import com.zipline.entity.user.User;
 import com.zipline.global.exception.custom.PermissionDeniedException;
 import com.zipline.global.exception.custom.UserNotFoundException;
 import com.zipline.global.exception.custom.customer.CustomerNotFoundException;
+import com.zipline.global.request.PageRequestDTO;
 import com.zipline.global.response.ApiResponse;
-import com.zipline.repository.CustomerRepository;
-import com.zipline.repository.UserRepository;
 import com.zipline.repository.agentProperty.AgentPropertyRepository;
 import com.zipline.repository.contract.CustomerContractRepository;
 import com.zipline.repository.counsel.CounselRepository;
+import com.zipline.repository.customer.CustomerRepository;
+import com.zipline.repository.user.UserRepository;
+import com.zipline.service.agentProperty.dto.response.AgentPropertyListResponseDTO;
+import com.zipline.service.contract.dto.response.ContractListResponseDTO;
+import com.zipline.service.counsel.dto.response.CounselListResponseDTO;
+import com.zipline.service.customer.dto.request.CustomerModifyRequestDTO;
+import com.zipline.service.customer.dto.request.CustomerRegisterRequestDTO;
+import com.zipline.service.customer.dto.response.CustomerDetailResponseDTO;
+import com.zipline.service.customer.dto.response.CustomerListResponseDTO;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
