+++ apiserver/service/src/main/java/com/zipline/service/customer/dto/response/CustomerDetailResponseDTO.java
@@ -1,4 +1,4 @@
-package com.zipline.dto;
+package com.zipline.service.customer.dto.response;
 
 import java.math.BigInteger;
